refactor to allow install via pip
