+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = create_multiresolution_meshes
+version = 0.0.1
+author = David Ackerman
+author_email = ackermand@janelia.hhmi.org
+description = convert single-scale (or a set of multi-scale) meshes to the neuroglancer multi-resolution mesh format
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/janelia-cosem/multiresolution-mesh-creator
+project_urls =
+    Bug Tracker = https://github.com/janelia-cosem/multiresolution-mesh-creator/issues
+classifiers =
+    Programming Language :: Python :: 3
+
+[options]
+packages = find:
+python_requires = >=3.7.11
+
+[options.entry_points]
+console_scripts =
+    create-multiresolution-meshes = multiresolution_mesh_creator.create_multiresolution_meshes:main
